+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/ADCWindow.java
@@ -12,23 +12,13 @@ package com.kaylerrenslow.armaDialogCreator.gui.fx.main;
 
 import com.kaylerrenslow.armaDialogCreator.data.DataKeys;
 import com.kaylerrenslow.armaDialogCreator.gui.canvas.api.ScreenDimension;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.popup.StagePopup;
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
-import com.kaylerrenslow.armaDialogCreator.main.lang.Lang;
-import com.kaylerrenslow.armaDialogCreator.util.BrowserUtil;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
-import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
-import javafx.geometry.Insets;
-import javafx.geometry.Orientation;
 import javafx.scene.Scene;
-import javafx.scene.control.Hyperlink;
-import javafx.scene.control.Label;
-import javafx.scene.control.Separator;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.VBox;
-import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 /**
@@ -40,9 +30,7 @@ public class ADCWindow {
 	private ADCCanvasView canvasView;
 	private ADCMenuBar mainMenuBar;
 	private boolean fullscreen = false;
-	private boolean betaShown = false;
-	
-	
+
 	public ADCWindow(Stage primaryStage) {
 		this.primaryStage = primaryStage;
 		primaryStage.fullScreenProperty().addListener(new ChangeListener<Boolean>() {
@@ -85,34 +73,6 @@ public class ADCWindow {
 		
 	public void show() {
 		this.primaryStage.show();
-		
-		if (!betaShown) {
-			new StagePopup<VBox>(primaryStage, new VBox(5), Lang.Popups.Beta.POPUP_TITLE) {
-				@Override
-				public void show() {
-					myStage.initModality(Modality.APPLICATION_MODAL);
-					
-					myRootElement.setPadding(new Insets(10));
-					myStage.setResizable(false);
-					final Label lblBody = new Label(Lang.Popups.Beta.BODY);
-					final Hyperlink hyperlink = new Hyperlink(Lang.Popups.Beta.REPORT_TO_LINK);
-					hyperlink.setOnAction(new EventHandler<ActionEvent>() {
-						@Override
-						public void handle(ActionEvent event) {
-							BrowserUtil.browse(Lang.Popups.Beta.REPORT_TO_LINK);
-						}
-					});
-					myRootElement.getChildren().addAll(lblBody, hyperlink, new Separator(Orientation.HORIZONTAL), getResponseFooter(false, true, false));
-					myStage.sizeToScene();
-					
-					btnOk.requestFocus();
-					btnOk.setDefaultButton(true);
-					
-					super.show();
-				}
-			}.show();
-			betaShown = true;
-		}
 	}
 	
 	public CanvasView getCanvasView() {