+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/popup/StageDialog.java
@@ -21,7 +21,8 @@ import javafx.stage.Stage;
 
 /**
  @author Kayler
- A wrapper class for {@link StagePopup} that is for dialogs.
+ A wrapper class for {@link StagePopup} that is for dialogs. The difference between a {@link StagePopup} and a {@link StageDialog} is that the {@link StageDialog} will always hault the thread when
+ shown and resume the thread when it is closed.
  Created on 08/20/2016. */
 public class StageDialog<T extends Parent> extends StagePopup<VBox>{